requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
pymysql==1.1.1
python-dotenv==1.0.1
requests==2.31.0